[FEATURE] 원두 및 캡슐 다중 필터링을 통한 검색 기능 추가

See merge request s09-bigdata-recom-sub2/S09P22A204!67

+++ backend/coffeeing/src/main/java/com/ssafy/coffeeing/modules/feed/dto/FeedDetailResponse.java
@@ -1,6 +1,6 @@
 package com.ssafy.coffeeing.modules.feed.dto;
 
-import com.ssafy.coffeeing.modules.tag.domain.Tag;
+import com.ssafy.coffeeing.modules.search.domain.Tag;
 
 import java.util.List;
 
@@ -10,7 +10,7 @@ public record FeedDetailResponse(
         String content,
         Tag tag,
         Long registerId,
-        Long likeCount,
+        Integer likeCount,
         String registerName,
         String registerProfileImg,
         Boolean isLike,
